+++ app/controllers/api/v1/static_invites_controller.rb
@@ -1,6 +1,7 @@
 module Api
   module V1
     class StaticInvitesController < Api::V1::BaseController
+      before_action :find_invite_creator, only: %i[index]
       before_action :find_static, only: %i[create]
       before_action :find_character, only: %i[create]
       before_action :check_static_member, only: %i[create]
@@ -9,6 +10,13 @@ module Api
       resource_description do
         short 'StaticInvite resources'
         formats ['json']
+      end
+
+      api :GET, '/v1/guild_invites.json', 'Get list of static_invites for static or character'
+      error code: 401, desc: 'Unauthorized'
+      def index
+        authorize! @from_static, with: StaticInvitePolicy, context: { static: @invite_creator, character: @invite_creator }
+        render json: (@invite_creator.is_a?(Static) ? @invite_creator.static_invites.includes(:character) : @invite_creator.static_invites.includes(:static)), status: 200
       end
 
       api :POST, '/v1/statics.json', 'Create static invite or member for guild members'
@@ -30,6 +38,20 @@ module Api
       end
 
       private
+
+      def find_invite_creator
+        if params[:static_id].present?
+          @invite_creator = Static.find_by(id: params[:static_id])
+          render_error(t('custom_errors.object_not_found'), 404) if @invite_creator.nil?
+          @from_static = 'true'
+        elsif params[:character_id].present?
+          @invite_creator = Character.where(user_id: Current.user.id).find_by(id: params[:character_id])
+          render_error(t('custom_errors.object_not_found'), 404) if @invite_creator.nil?
+          @from_static = 'false'
+        else
+          render_error('Static ID or Character ID must be presented', 400)
+        end
+      end
 
       def find_static
         @static = Static.find_by(id: params[:static_invite][:static_id])