+++ db/migrate/20191209205411_add_from_static_to_static_invites.rb
@@ -1,0 +1,5 @@
+class AddFromStaticToStaticInvites < ActiveRecord::Migration[5.2]
+  def change
+    add_column :static_invites, :from_static, :boolean, null: false, default: true
+  end
+end