+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2019_12_09_195453) do
+ActiveRecord::Schema.define(version: 2019_12_09_205411) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "pg_trgm"
@@ -258,6 +258,7 @@ ActiveRecord::Schema.define(version: 2019_12_09_195453) do
     t.integer "status", default: 0, null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
+    t.boolean "from_static", default: true, null: false
     t.index ["static_id", "character_id"], name: "index_static_invites_on_static_id_and_character_id", unique: true
     t.index ["status"], name: "index_static_invites_on_status"
   end