+++ spec/api/v1/static_invites_controller_spec.rb
@@ -1,4 +1,110 @@
 RSpec.describe 'StaticInvites API' do
+  describe 'GET#index' do
+    let!(:guild) { create :guild }
+    let!(:static) { create :static, staticable: guild }
+
+    it_behaves_like 'API auth without token'
+    it_behaves_like 'API auth with invalid token'
+    it_behaves_like 'API auth unconfirmed'
+
+    context 'with valid user token in params' do
+      let!(:user) { create :user }
+      let(:access_token) { JwtService.new.json_response(user: user)[:access_token] }
+
+      context 'without static or character param' do
+        before { get '/api/v1/static_invites.json', params: { access_token: access_token } }
+
+        it 'returns status 400' do
+          expect(response.status).to eq 400
+        end
+
+        it 'and returns error message' do
+          expect(JSON.parse(response.body)).to eq('error' => 'Static ID or Character ID must be presented')
+        end
+      end
+
+      context 'with static param' do
+        context 'for unexisted static' do
+          before { get '/api/v1/static_invites.json', params: { access_token: access_token, static_id: 'unexisted' } }
+
+          it 'returns status 404' do
+            expect(response.status).to eq 404
+          end
+
+          it 'and returns error message' do
+            expect(JSON.parse(response.body)).to eq('error' => 'Object is not found')
+          end
+        end
+
+        context 'for existed static' do
+          let!(:character) { create :character, user: user }
+          let!(:static) { create :static, staticable: character }
+          let!(:static_invite) { create :static_invite, static: static }
+          before { get '/api/v1/static_invites.json', params: { access_token: access_token, static_id: static.id } }
+
+          it 'returns status 200' do
+            expect(response.status).to eq 200
+          end
+
+          %w[id static_name character from_static status].each do |attr|
+            it "and contains static_invite #{attr}" do
+              expect(response.body).to have_json_path("static_invites/0/#{attr}")
+            end
+          end
+        end
+      end
+
+      context 'with character param' do
+        context 'for unexisted character' do
+          before { get '/api/v1/static_invites.json', params: { access_token: access_token, character_id: 'unexisted' } }
+
+          it 'returns status 404' do
+            expect(response.status).to eq 404
+          end
+
+          it 'and returns error message' do
+            expect(JSON.parse(response.body)).to eq('error' => 'Object is not found')
+          end
+        end
+
+        context 'for existed character' do
+          context 'for not user character' do
+            let!(:character) { create :character }
+            before { get '/api/v1/static_invites.json', params: { access_token: access_token, character_id: character.id } }
+
+            it 'returns status 404' do
+              expect(response.status).to eq 404
+            end
+
+            it 'and returns error message' do
+              expect(JSON.parse(response.body)).to eq('error' => 'Object is not found')
+            end
+          end
+
+          context 'for valid character' do
+            let!(:character) { create :character, user: user }
+            let!(:static_invite) { create :static_invite, static: static, character: character, from_static: true }
+            before { get '/api/v1/static_invites.json', params: { access_token: access_token, character_id: character.id } }
+
+            it 'returns status 200' do
+              expect(response.status).to eq 200
+            end
+
+            %w[id static_name character from_static status].each do |attr|
+              it "and contains static_invite #{attr}" do
+                expect(response.body).to have_json_path("static_invites/0/#{attr}")
+              end
+            end
+          end
+        end
+      end
+    end
+
+    def do_request(headers = {})
+      get '/api/v1/static_invites.json', params: { static_id: static.id }, headers: headers
+    end
+  end
+
   describe 'POST#create' do
     it_behaves_like 'API auth without token'
     it_behaves_like 'API auth with invalid token'